+++ src/main/java/frc/robot/Vision/AprilTag.java
@@ -1,4 +1,4 @@
-package frc.robot.Vision;
+package frc.robot.vision;
 
 public class AprilTag {
